discord-bots/requirements.txt
discord.py>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
httpx>=0.27.0
